move selector to selectors file
